Add method to check all images in the page exist #17
This commit bumps the minimum Perl version to 5.10 in order to
accomodate new required dependencies.

This commit also bumps the required version of HTTP::Message to 6.29 to
fix an issue with the canonical URL cache that this implementation ran
into.

The gif included in this commit is originally taken from
https://github.com/mathiasbynens/small.